AccountManagerPlugin: Ensure sensible browser cookie lifetime setting, refs #8927, #9088, #9095, #9099 and #9547.
I think this is now the most intuitive setting of default cookie lifetime:
`auth_cookie_lifetime` from section `[trac]` gets overwritten with !AcctMgr
default (30 d) as long as it's found equal to the Trac default (0).

Remember, that the !AcctMgr feature still has to be switched on with the
boolean option `persistent_sessions`, that defaults to `False`, if unset.