Refs #10194: Implemented a fix for the `IntegrityError` when importing tickets with the TicketImportPlugin and not defining a changetime in the dataset (or in some cases of using the same changetime for multiple tickets).
The `ticket id`, `time` and `field` columns must be unique in the `ticket_change` table. Importing a set of tickets that are all assigned a changetime of "now" will result in an `IntegrityError` whenever two or more ticket share a relation with a single other ticket.

This feels like more of a workaround than a solution, but it will hopefully do for now.